requirements.txt
django
djangorestframework
requests
django-cors-headers
tensorflow
pillow
openai
python-dotenv